[MIGRATIONS] More sanity checks

+++ core/migrations/Migration20141120215253ComHubgraph.php
@@ -23,34 +23,37 @@ class Migration20141120215253ComHubgraph extends Base
 		);
 		$this->addComponentEntry('Hubgraph', 'com_hubgraph', 1, $params, false);
 
-		// Look for multiple entries
-		$query = "SELECT `extension_id` FROM `#__extensions` WHERE `element` = 'com_hubgraph' ORDER BY `extension_id` ASC";
-		$this->db->setQuery($query);
-		$ids = $this->db->loadColumn();
-
-		if ($ids && count($ids) > 1)
+		if ($this->db->tableExists('#__extensions'))
 		{
-			unset($ids[0]);
+			// Look for multiple entries
+			$query = "SELECT `extension_id` FROM `#__extensions` WHERE `element` = 'com_hubgraph' ORDER BY `extension_id` ASC";
+			$this->db->setQuery($query);
+			$ids = $this->db->loadColumn();
 
-			foreach ($ids as $id)
+			if ($ids && count($ids) > 1)
 			{
-				$query = "DELETE FROM `#__extensions` WHERE `extension_id` = " . (int)$id;
+				unset($ids[0]);
+
+				foreach ($ids as $id)
+				{
+					$query = "DELETE FROM `#__extensions` WHERE `extension_id` = " . (int)$id;
+					$this->db->setQuery($query);
+					$this->db->query();
+				}
+			}
+
+			// Look for non-json params
+			$params = $this->getParams('com_hubgraph', true);
+
+			if (is_null(json_decode($params)))
+			{
+				$object = unserialize($params);
+				$params = $object->settings();
+
+				$query = "UPDATE `#__extensions` SET `params` = " . $this->db->quote(json_encode($params)) . " WHERE `element` = 'com_hubgraph'";
 				$this->db->setQuery($query);
 				$this->db->query();
 			}
-		}
-
-		// Look for non-json params
-		$params = $this->getParams('com_hubgraph', true);
-
-		if (is_null(json_decode($params)))
-		{
-			$object = unserialize($params);
-			$params = $object->settings();
-
-			$query = "UPDATE `#__extensions` SET `params` = " . $this->db->quote(json_encode($params)) . " WHERE `element` = 'com_hubgraph'";
-			$this->db->setQuery($query);
-			$this->db->query();
 		}
 	}
 }
